[server,client, Connectors] Use a generic object for api keys and user entered params.
## Motivation

Instead of using API Key as a concept we decided to use a more generic approach that saves all the params that a user can enter in our auth step.

## Change description

We create the auth type `USER_PROVIDED_PARAMS`
We use the above to save the data in the credentials in an object (we were already doing this but only for apiKey).
We include this new extra property to the places we were using apiKeys.

We deprecate API_KEY (I don't think removing it is a good idea as we will run into issues with migrations and such).
If needed we can do the above in a diff issue later on.


## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High