[server] Auth parser for connectors.
## Motivation

Connectors like Wordpress need extra processing before saving the credentials that the user has given us.

## Change description

We implement a new interface AuthParser which has a method to parse the credentials and extras needed for a connector that runs before we save a new connection.
If the connector doesn't have one we can just default to what the user gave us.


## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High