[client, server] Doing renaming to Scratch
## Motivation

We still had "Scratchpaper" all over the app, especially on the home screen. Now that we're switching to Scratch, I wanted to do a quick cleanup.

## Change description

Renaming things to Scratch. I went through the client and the server code to find references to Scratchpaper that could easily be changed to Scratch without affecting any functionality.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High