[agent] clean up fetch tool histories
## Motivation

When we have data fetching tools chat history builds up quickly. The snapshot preview is part of the system instructions so it gets droped and regenerated with each prompt. The tool call responses, however, remain. Unless we manually clean them up. 
I think context management will be the core of Scratch. This is just scratching the surface (pun intended). Next version will involve the LLM making a suggestion for what needs to be cleaned up. For example we might fetch 100 records with the intention to query the llm about them a few times. In that case we should not immediately clean the record data up. Otherwise the agent will have to start each of the next few prompts with refetching the same records. 

## Change description

- extract the record `data` returned by fetch tool calls into its own json field
- a bunch of other fields can be returned too
- history processing cleans up all but the last ModelRequest message
- for data fetch tools this means - remove the data
- for other tools it means - make sure it is not longer than 1000 characters

## TODO:
get the AI to suggest what to be cleaned up


## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High