[server, database]: Make knex clients use a secure connection without verifying the server cert
## Motivation

We can't verify the server cert of the CloudSQL server (yet), but at least using a secure connection is better than nothing.

## Change description

Make a reusable method on DbService for getting a Knex client that configures the client based on the connection string to use SSL but not verify the server cert.


## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High