[client/server, snapshots] Snapshot tables cleanup 1: Scope down column settings
## Motivation

I want to be able to make a blank Snapshot, which has led me down a rabbit hole. I need to complete the migration from Snapshot to SnapshotTable, there’s a few things left over.

## Change description

Change #1: columnContexts. This is still a map from tableId->columnId->settings. Now that it’s owned by one table, it needs to be just columnId->settings. Along the way, i’ve renamed it to `columnSettings`.

It’s still not consistently coming from the same column in the DB, i’ll do that next.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High