[server,client] Use connector list from feature flags.
## Motivation

We need to manage connector with flags.

## Change description

We add a new json flag to the server and then we add the CONNECTOR_LIST flag to the client+server, keep this as a user flag since we will probably want to give access to users to certain connectors in the future as we have test users.


## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High