[client] Move mantine style sheet to root layout
## Motivation

Custom styles and built-in styles were being applied in the wrong order (and not even consistently so)

## Change description

Mantine's `'@mantine/core/styles.css'` needs to be imported before any theme styles or custom styles. I moved it back to the `app/layout.tsx`. There might be more to fix-up, but we'll deal with that as we find it.
Menu's default widget gets custom styles that actually render now!

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High