[cleanup, client] remove AI custom connector builder demo
## Motivation

This experimental tool has not been kept up to date and is no longer used by anyone, but is leaving a lot of code lying around. Time to remove it from the codebase — we can rebuild it later when we have a better use case for it.

## Change description

This MR just removes the connector demo from the `client` code

* Remove the CUSTOM connector service type
* remove other custom connector support code
* remove nav elements
* Remove the glide package
* remove the builder page

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High