[infra] refactor SnapshotTable id/wsid/slug/name
## Change description

The main identifier for our `SnapshotTable`s was `snapshotTable.tableSpec.id.wsId`. 
This was strange for a number of reasons and had some side effects. 
This is a massive refactor that makes snapshotTable.id the main identifier used across server/client/agent and creates a migration utility to standardize the naming conventions for old and new tables. 

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High