[server] ignoring suggestions for an update to a readonly field
## Motivation

At the moment these are accepted and when the agent gets confused and proposes an update to a readonly field (ex: last_edited), accepting the change fails. We cannot use "accept all" but need to go click 1 by 1 on each field, or figure out which field is broken and reject it first.

## Change description

The server ignores readonly fields in a suggestion and throws an error in accept operations. 

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High