[server] Track actions in a new entity in the db.
## Motivation

We need a way to know how many publishes has a user done so we can limit that amount based on the subscription the user has.
For the above we add a new entity in the DB called Action which can contain any type of action we want and we can query it and find out exactly how many times it happened for a org/user. We can use this for multiple type of actions we have a metadata field where we can store action specific information.

## Change description

* [server] Track actions in a new entity in the db.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High