[client] html diff
## Motivation

Make a suggestion in a rich text field from Webflow (now that it's sending HTML) and you'll see the diff viewers all render it. It looks like it's dangerously setting inner html, which is probably dangerous for agent and remote data anyway.

## Change description

Converted the diff to always render the raw text instead of dangerously setting the html content of the diff spams. Diff rendering was the main issue, but now we don't need to worry about dangerous html too.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High