[client, server] Unbreak credentials creation
## Motivation

Credentials creation doesn’t work because the DTO isn’t being instantiated, so the validations get mad at missing decorations.

## Change description

Instantiate the DTOs. At the same time, i converted it to use a form, so the DTO isn’t the same object being maniuplated by the UI, which gives validation and defaults.

I had to add a few properties for tokenUsageWarningLimit, since while i was testing, it wasn’t working and i wasn’t sure if i broke it or not.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High