[client, server] Add models costs and history calculation to ai chat panel.
## Motivation

We need to calculate the cost of sessions.
This might need some improvements to decrease the amount of processing we do, but for now it should be alright.
We fetch model pricings from openrouter then the aichat panel component calculates with the history the amount of dollars we are spending in a session.
## Change description

* [client, server] Add models costs and history calculation to ai chat panel.

## Checklists

- [x] Security impact of change has been considered
- Data Incident Risk: 
    - [x] None
    - [ ] Low
    - [ ] High